requirements.txt
torch==2.2.2
torchvision==0.17.2
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
matplotlib==3.8.4
seaborn==0.13.2
pillow==10.3.0
mlflow==2.12.1
protobuf==4.25.3
setuptools==70.0.0
pathspec==0.11.2
dvc[all]==3.51.2
pyyaml==6.0.1
tqdm==4.66.4
pytest==8.2.0
fastapi==0.111.0
uvicorn==0.30.0
python-multipart==0.0.9